requirements.txt
disnake==2.8.1
python-dotenv==1.0.0
giphy-client==1.0.0